+++ semi/src/semi/beans/board/BA_Board/BA_BoardDto.java
@@ -1,4 +1,4 @@
-package semi.beans.BA_Board;
+package semi.beans.board.BA_Board;
 
 public class BA_BoardDto {
 	private int board_no;
