+++ semi/src/semi/beans/board/BoardDao.java
@@ -1,4 +1,4 @@
-package semi.beans;
+package semi.beans.board;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
