+++ semi/src/semi/beans/board/BoardReplyDto.java
@@ -1,4 +1,4 @@
-package semi.beans;
+package semi.beans.board;
 
 public class BoardReplyDto {
 	int reply_no;
