+++ semi/src/semi/beans/board/member/MemberDto.java
@@ -1,4 +1,4 @@
-package beans;
+package semi.beans.board.member;
 
 public class MemberDto {
