+++ semi/src/semi/servlet/ba_Board/BA_FileDownloadServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package semi.servlet.ba_Board;
 
 import java.io.File;
 import java.io.IOException;
@@ -12,8 +12,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.io.FileUtils;
 
-import semi.beans.BA_Board.BA_FileDao;
-import semi.beans.BA_Board.BA_FileDto;
+import semi.beans.board.BA_Board.BA_FileDao;
+import semi.beans.board.BA_Board.BA_FileDto;
 @WebServlet(urlPatterns = "/ba_board/download.do")
 public class BA_FileDownloadServlet extends HttpServlet {
 
@@ -25,9 +25,18 @@ public class BA_FileDownloadServlet extends HttpServlet {
 		BA_FileDao fdao=new BA_FileDao();
 		BA_FileDto fdto=fdao.get(no);
 		
-		File target=new File("D:/upload/kh21/ba_board",fdto.getSavename());
+		String path="D:/upload/kh21/ba_board";
+		
+		File target=new File(path,fdto.getSavename());
 		byte[] data=FileUtils.readFileToByteArray(target);
 		
+		if(!target.exists()) {
+			target.mkdir();//폴더 생성
+			System.out.println("생성완료");
+		}
+		else {
+			System.out.println("이미 폴더 있음");
+		}	
 		resp.setHeader("Content-Type", "application/octet-srtream; charset=UTF-8");
 		resp.setHeader("Content-Disposition", "attachment; filename=\""+URLEncoder.encode(fdto.getUploadname(),"UTF-8")+"\"");
 		resp.setHeader("Content-Length", String.valueOf(fdto.getFilesize()));
