+++ semi/src/semi/servlet/board/BoardEditServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package semi.servlet.board;
 
 import java.io.IOException;
 
@@ -8,22 +8,24 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns ="/member/check.do")
-public class MemberCheckServlet extends HttpServlet{
+import semi.beans.board.BoardTextDao;
+import semi.beans.board.BoardTextDto;
 
+@WebServlet(urlPatterns="/board/boardedit.do")
+public class BoardEditServlet extends HttpServlet{
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-	
 		try {
-		//	수신
+			BoardTextDto boardtextdto = new BoardTextDto();
+			boardtextdto.setNo(Integer.parseInt(req.getParameter("no")));
 			
 			
-			
+			BoardTextDao boardtextdao = new BoardTextDao();
+			resp.sendRedirect("searchResult.jsp?keyword="+req.getParameter("keyword"));
 		}
 		catch(Exception e) {
 			e.printStackTrace();
 			resp.sendError(500);
 		}
 	}
-	
 }
