+++ semi/src/semi/servlet/member/MemberCheckServlet.java
@@ -1,4 +1,4 @@
-package semi.servlet;
+package semi.servlet.member;
 
 import java.io.IOException;
 
@@ -8,24 +8,22 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import semi.beans.BoardTextDao;
-import semi.beans.BoardTextDto;
+@WebServlet(urlPatterns ="/member/check.do")
+public class MemberCheckServlet extends HttpServlet{
 
-@WebServlet(urlPatterns="/board/boardedit.do")
-public class BoardEditServlet extends HttpServlet{
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+	
 		try {
-			BoardTextDto boardtextdto = new BoardTextDto();
-			boardtextdto.setNo(Integer.parseInt(req.getParameter("no")));
+		//	수신
 			
 			
-			BoardTextDao boardtextdao = new BoardTextDao();
-			resp.sendRedirect("searchResult.jsp?keyword="+req.getParameter("keyword"));
+			
 		}
 		catch(Exception e) {
 			e.printStackTrace();
 			resp.sendError(500);
 		}
 	}
+	
 }
