+++ semi/src/semi/servlet/member/MemberFindServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package semi.servlet.member;
 
 import java.io.IOException;
 
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import beans.MemberDao;
+import semi.beans.board.member.MemberDao;
 
 @WebServlet(urlPatterns="/member/find.do")
 public class MemberFindServlet extends HttpServlet{
