+++ semi/src/semi/servlet/member/MemberLogoutServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package semi.servlet.member;
 
 import java.io.IOException;
